+++ packages/expo-updates/android/src/main/java/expo/modules/updates/launcher/DatabaseLauncher.kt
@@ -71,7 +71,7 @@ class DatabaseLauncher(
 
     launchedUpdate = getLaunchableUpdate(database, context)
     if (launchedUpdate == null) {
-      this.callback!!.onFailure(Exception("No launchable update was found. If this is a bare workflow app, make sure you have configured expo-updates correctly in android/app/build.gradle."))
+      this.callback!!.onFailure(Exception("No launchable update was found. If this is a generic app, ensure expo-updates is configured correctly."))
       return
     }
 
@@ -83,9 +83,14 @@ class DatabaseLauncher(
 
     // verify that we have all assets on disk
     // according to the database, we should, but something could have gone wrong on disk
-    val launchAsset = database.updateDao().loadLaunchAsset(launchedUpdate!!.id)
+    val launchAsset = database.updateDao().loadLaunchAssetForUpdate(launchedUpdate!!.id)
+    if (launchAsset == null) {
+      this.callback!!.onFailure(Exception("Launch asset not found for update; this should never happen. Debug info: ${launchedUpdate!!.debugInfo()}"))
+      return
+    }
+
     if (launchAsset.relativePath == null) {
-      throw AssertionError("Launch Asset relativePath should not be null")
+      this.callback!!.onFailure(Exception("Launch asset relative path should not be null. Debug info: ${launchedUpdate!!.debugInfo()}"))
     }
 
     val launchAssetFile = ensureAssetExists(launchAsset, database, context)
@@ -113,7 +118,7 @@ class DatabaseLauncher(
 
     if (assetsToDownload == 0) {
       if (this.launchAssetFile == null) {
-        this.callback!!.onFailure(Exception("mLaunchAssetFile was immediately null; this should never happen"))
+        this.callback!!.onFailure(Exception("Launch asset file was null with no assets to download reported; this should never happen. Debug info: ${launchedUpdate!!.debugInfo()}"))
       } else {
         this.callback!!.onSuccess()
       }